Move user related items into /modules/user; update imports
