+++ src/modules/user/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './change-password.input'
+export * from './update-user.input'
