+++ src/modules/user/user.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common'
-import { PasswordService } from 'src/services/password.service'
-import { PrismaService } from 'src/services/prisma.service'
-import { UserService } from 'src/services/user.service'
+import { PrismaService } from 'src/services'
 
+import { PasswordService } from '../auth/services'
 import { UserResolver } from './user.resolver'
+import { UserService } from './user.service'
 
 @Module({
   providers: [UserResolver, UserService, PasswordService, PrismaService]
