--- src/resolvers/user/dto/change-password.input.ts
@@ -1,15 +1,0 @@
-import { InputType, Field } from '@nestjs/graphql'
-import { IsNotEmpty, MinLength } from 'class-validator'
-
-@InputType()
-export class ChangePasswordInput {
-  @Field()
-  @IsNotEmpty()
-  @MinLength(8)
-  oldPassword: string
-
-  @Field()
-  @IsNotEmpty()
-  @MinLength(8)
-  newPassword: string
-}
